first commit for timestamp service as a lib
