--- user_services/timestamp/playbooks/get_timestamp_service_status.yaml
@@ -1,10 +1,0 @@
-- name: Start timesatmp service
-  hosts: Experiment_Nodes
-  tasks:
-    - name: Get Service Status
-      systemd:
-        name: timestamp.service
-      register: timestamp_service_status
-      
-    - debug:
-        var: timestamp_service_status.status.ActiveState
