--- user_services/timestamp/playbooks/stop_timestamp_service.yaml
@@ -1,9 +1,0 @@
-- name: Stop timesatmp service
-  hosts: Experiment_Nodes
-  tasks:
-    - name: stop timestamp.service
-      systemd:
-        name: timestamp.service
-        state: stopped
-        enabled: no
-
