--- user_services/timestamp/timestamp/service_files/timestamp.service
@@ -1,9 +1,0 @@
-[Unit]
-Description=Timestamp Service
-After=multi-user.target
-[Service]
-Type=simple
-Restart=always
-ExecStart=/usr/bin/python3 /home/mfuser/services/timestamp/service_files/timestampservice.py
-[Install]
-WantedBy=multi-user.target
